make some chonges

+++ controllers/productController.js
@@ -22,37 +22,57 @@ exports.show_products = (request, response, next) => {
 };
 
 exports.add_product = (request, response, next) => {
+  if(request.role == "admin" || request.role == "seller"){
+    let object = new Products({
+      name: request.body.name,
+      price: request.body.price,
+      brand: request.body.brand,
+      category_id: request.body.category_id,
+      discount: request.body.discount,
+      reviews: request.body.reviews,
+      description: request.body.description,
+      images: request.body.images,
+      properties: request.body.properties,
+      quantity: request.body.quantity,
+      seller: request.body.seller,
+    });
+    object
+      .save()
+      .then((data) => {
+        response.status(201).json({ message: "added", data });
+      })
+      .catch((error) => next(error));
+  }
+  else{
+    response.status(404).json({ mesg: "You should be admin or seller to add product" });
+  }
   
-  let object = new Products({
-    name: request.body.name,
-    price: request.body.price,
-    brand: request.body.brand,
-    category_id: request.body.category_id,
-    discount: request.body.discount,
-    reviews: request.body.reviews,
-    description: request.body.description,
-    images: request.body.images,
-    properties: request.body.properties,
-    quantity: request.body.quantity,
-    seller: request.body.seller,
-  });
-  object
-    .save()
-    .then((data) => {
-      response.status(201).json({ message: "added", data });
-    })
-    .catch((error) => next(error));
   // });
 };
 
 exports.delete_product = (request, response, next) => {
-  Products.findByIdAndDelete({ _id: request.body.id })
+  if(request.role == "admin"){
+    Products.findByIdAndDelete({ _id: request.body.id })
     .then((data) => {
       response.status(201).json({ message: "deleted", data });
     })
     .catch((error) => {
       next(error);
     });
+  }
+  else{
+    Products.findById({_id:request.body.id})
+      .then((data) =>{
+        if(request.role == data.seller.useID){
+          response.status(201).json({ message: "the same", data });
+        }
+        else{
+          response.status(404).json({ message: "You dont owne this product " });
+        }
+      })
+  }
+
+  
 };
 
 exports.update_stock = (request, response, next) => {
